+++ docs/dev/CHANGELOG.md
@@ -6,6 +6,7 @@ Released on: ???
 
 - Added Expertise kills counter to the Items settings section, showing exact expertise counter on fishing rods [disabled by default].
 - Added option to set a keybind for bulk resetting multiple trackers' data, to quickly reset current fishing session. You can also choose what trackers to include.
+- Added option "Do not activate when wearing Trophy armor" for Fishing profit tracker [disabled by default].
 
 ## Bugfixes